requirements.txt
Flask-PyMongo==2.3.0
bs4
requests
# pandas==0.23.0
